v1.0 Delivery.
Acceptance Criteria:

1) Record a square video with built-in camera.
2) Limit recording to 30s or less.
3) Save recorded video to device. (mp4)
4) Plays back a preview when finished recording.

Additionally:
Added a recorded video gallery as a placeholder.
Added a visual representation of the 30 second timer to the video preview texture.
Added logic to lock orientation while recording.
Added a video gallery/adpater (RecyclerView) as a placeholder for additional tasks. Please note: This gallery is for placement only!

Out of time here, but would have liked to:
	a)obtain video still shot (thumbnail) of each recorded video
	b)populate video gallery with these thumbnails
	c)and use the thumbnails as buttons to launch recorded videos

TODO:
Watch some football and enjoy the rest of my weekend. Thanks for stopping by!